refactor(dashboard): reorganize components into pages directory
- Move Dashboard and All Files components to pages directory
- Add FileSkeleton component for loading states
- Update imports and file structure for better organization
