+++ components/DashboardSkeleton.tsx
@@ -5,8 +5,8 @@ export default function DashboardSkeleton() {
     <div className="flex flex-col gap-6">
       {/* Header Section */}
       <div className="flex flex-col gap-2">
-        <Skeleton className="h-8 w-48" />
-        <Skeleton className="h-5 w-96" />
+        <Skeleton className="h-7 w-48" />
+        <Skeleton className="h-4 w-96" />
       </div>
 
       {/* Stats Cards */}
